Parallelize uploads.
Restructure upload preparation code and data structures.
For folder uploads, there are now three distinct phases in the
upload pipeline. First, all files in the upload are prepared
(compressed/encrypted/erasure coded) concurrently. Second,
the main thread allocates storage for the files as each finishes
preparation. Finally, a separate worker thread is used for each
provider involved in the uploads, and blocks from the prepared
uploads are fed to the appropriate worker for transport to their
destination.